Added SDG_Min/SDG_Max/SDG_Clamp macros, made Math templates more user-friendly

+++ src/SDG/Math/MathEx.h
@@ -28,8 +28,8 @@ namespace SDG::Math
      */
     inline Vector2 Trajectory(float degrees, float length)
     {
-        return Vector2((float)TrajectoryX(degrees, length),
-                       (float)TrajectoryY(degrees, length));
+        return Vector2(TrajectoryX(degrees, length),
+                       TrajectoryY(degrees, length));
     }
 
     /// Constrains a value between two Vectors, inclusively on both ends.
